Move default merge strategy to own module
